deferred_rollback: better handling of invalid nested rollbacks
Currently, there is a messy case when savepoints aren't enabled
and a user creates nested transactions and then calls 'rollback'.
DBIC throws an exception in hopes that it propogates up the stack
to the top-most transaction where it can be rolled back properly,
but there is no guarantee that the user doesn't just catch the
exception and then accidentally commit unwanted changes by
exiting a txn_do block cleanly.

This change is a proposed fix that sets a "deferred rollback"
condition on the Storage object such that attempting any other
transaction-related operation will fail until the top-level
transaction has been rolled back.  It also makes the BlockRunner
aware of deferred rollbacks so that it performs a rollback
instead of commit when a block exits cleanly.

It also takes the somewhat bold step of stopping BlockRunner from
starting new blocks at all until the condition is resolved...

I stopped short of reaching into the individual Storage engines
to stop them from executing further statements, though that is
theoretically what we want to do.