Test case for caught errors from nested rollback
There are several ways that a nested rollback can cause
un-intended behavior, but this is one I feel is pretty realistic.
(In fact, it probably already happens in the wild somewhere.)

In this case, if a subroutine wraps some code in a transaction,
then catches errors from the code expecting the transaction to be
rolled back, then performs some of its own database work without
carefully inspecting the exception it caught, the whole thing can
end up geting committed if it occurs within an inner transaction.