Add SQLite-based versions of the live Oracle id-shortener test
Note that this is *NOT* Oracle specific: there is way more use of
_quote()-as-a-filter in the wild, as it is a perfect hook point for last-stage
mangling, and has been recommended for this to various parties.

In the process of writing this test also find a subtle inconsistency in the
bindtype inference codepath due to silly eager-encoding of identifiers.

This will sadly have to wait until master-time :/