Provide an additional avenue for SQLAC-compatible maker tests
An extra envvar DBICDEVREL_SWAPOUT_SQLAC_WITH takes a classname, and will
take effect IFF the current version of DBIC is a devrel. This allows one
to smoke DBIC dependents by simply installing a recent dev version of DBIC
and setting the environment variable in question.

The design is motivated by the need to confine usage of this mechanism
exclusively to its intended purposes of testing, preventing production
environments which likely use a "stable" DBIC version from being affected.