Fix test failures with SQlite 3.37.0
In prior versions DBD::SQLite would preserve the case of data types
as provided to the CREATE TABLE statement. This is no longer the case
in newer versions, where INTEGERs seem to be normalized to upper case.

Normalizing within DBIx::Class::Storage::DBI is not a viable approach
as it will destabilize too many applications relying on roundtripping
via ::Schema:Loader and the like.

Just work around in tests with case-insensitive matches :/